poc_pagos/requirements.txt
fastapi
uvicorn
pika
kafka-python
python-dotenv
pydantic